backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiofiles>=24.1.0",
    "fastapi>=0.116.1",
    "feedparser>=6.0.12",
    "httpx>=0.28.1",
    "openai>=1.54.0",
    "pydantic>=2.11.9",
    "pydub>=0.25.1",
    "python-dotenv>=1.0.0",
    "sse-starlette>=2.1.3",
    "uvicorn>=0.35.0",
]
